Move stash() and clear_repo() to git_helper
