+++ tests/ci/cherry_pick.py
@@ -26,19 +26,16 @@ Cherry-pick stage:
 import argparse
 import logging
 import os
-from contextlib import contextmanager
 from datetime import date, datetime, timedelta
 from pathlib import Path
 from subprocess import CalledProcessError
 from typing import List, Optional
 
-import __main__
-
 from ci_buddy import CIBuddy
 from ci_config import Labels
 from env_helper import IS_CI, TEMP_PATH
 from get_robot_token import get_best_robot_token
-from git_helper import GIT_PREFIX, git_runner, is_shallow
+from git_helper import GIT_PREFIX, git_runner, is_shallow, stash
 from github_helper import GitHub, PullRequest, PullRequests, Repository
 from ssh import SSHKey
 
@@ -569,7 +566,10 @@ class Backport:
 
 
 def parse_args():
-    parser = argparse.ArgumentParser("Create cherry-pick and backport PRs")
+    parser = argparse.ArgumentParser(
+        "Create cherry-pick and backport PRs",
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+    )
     parser.add_argument("--token", help="github token, if not set, used from smm")
     parser.add_argument(
         "--repo", default="ClickHouse/ClickHouse", help="repo owner/name"
@@ -594,39 +594,6 @@ def parse_args():
         help="add debug logging for git_helper and github_helper",
     )
     return parser.parse_args()
-
-
-@contextmanager
-def clear_repo():
-    def ref():
-        return git_runner("git branch --show-current") or git_runner(
-            "git rev-parse HEAD"
-        )
-
-    orig_ref = ref()
-    try:
-        yield
-    finally:
-        current_ref = ref()
-        if orig_ref != current_ref:
-            git_runner(f"git checkout -f {orig_ref}")
-
-
-@contextmanager
-def stash():
-    # diff.ignoreSubmodules=all don't show changed submodules
-    need_stash = bool(git_runner("git -c diff.ignoreSubmodules=all diff HEAD"))
-    if need_stash:
-        script = (
-            __main__.__file__ if hasattr(__main__, "__file__") else "unknown script"
-        )
-        git_runner(f"git stash push --no-keep-index -m 'running {script}'")
-    try:
-        with clear_repo():
-            yield
-    finally:
-        if need_stash:
-            git_runner("git stash pop")
 
 
 def main():
